WIP: Rework the task list layout
This commit is currently WIP. It implements a rework of the task list
UI in the story view, aimed at reducing wrapping of fields and making
space for the various bits we want to add. The new fields (the add to
worklist button and the gerrit link) are only mock-ups at the moment.

Change-Id: I8694c906c0a5f3567331574ec3a516dec586414a