Allow teams to be added to the permission list for private stories
This commit adds a way to add teams to the permission list for private
stories. This makes it easy to use for stories that should only be
visible to a given set of people; the members of that set can now be
managed by the teams UI rather than on a per-story basis.

Change-Id: I7d3adde5d6cf9b4d5b7e63df06d2266f7b824666
Depends-On: Idb7f881c4d772d373b583ecc2f59231be8f7f01b