Put task notes in a box
This distinguishes task notes from other elements visually, so they're
a little easier to scan. It also aligns the 'add notes' and
'edit'(notes) buttons, so that the user is not confused, as they will
expect these to be in the same place. Finally, it puts them both nearer
the task notes, so that it's obvious what the edit button edits.

Change-Id: I152a8dc164c9173b52f25b000eb65af9a1b7c57f