Give certain lists a border
Sometimes borders are useful for lists as they make it easier
to look at the right thing quickly. Other times, they just make
things look crowded. This is an attempt to neaten up the dashboard
and subscription pages, and the tasks list in stories. I'm not
sure about it; maybe we don't want this on all these pages (or any!);
maybe it's an improvement and I'm just used to StoryBoard's current
appearance. So, comment away!

Change-Id: I15e0efc429c3b181e32c2686b5b407f14ead86d7