Order projects in task list alphabetically
Previously, in the task list on the story detail view,
projects just displayed in the order that they loaded. This meant
that groups of tasks might display in a different order each time
a story was viewed, which was confusing (and would make it harder
for a user to find tasks for the project they were
interested in for any given story, as the projects would not be in a
guessable order). This patch orders the projects alphabetically.

Change-Id: I69e4a82e53de9ae5d8cefbe9844f08a3ae26e135