Converting less to scss
In process of converting less to scss in storyboard-webclient.
Original purpose for using less is due to sass previously requiring
ruby-compass which could not be easily installed with npm. This is
no longer the case and sass has far greater support than less, is
easier to debug, and supports functions and named arguments for ease
of coding and ongoing code maintenance.

Debugging should also be easier as sourcemaps are now enabled for scss
so that in browser tools you can see the scss file and line that the css
relates to as opposed to the compiled css file.

Change-Id: I22fcbf4504d7756cd7ece41b0eca12985316db6b