Replace marked.js with markdown-it.js
The markdown parser used in StoryBoard (marked.js) is buggy and
mostly unmaintained. This commit replaces it with a better library
which is actually maintained.

There is a small change to functionality incurred with this change,
we no longer automatically attempt to highlight every code block by
guessing the language. Instead only fenced code blocks are
highlighted, and then only if the language is specified. For example:

    ``` python
    def foo():
        return 'foo'
    ```

Change-Id: I8ee562e3e97ab6f0d09a21c27907d1a0959aceb1