Add a "New Story" view
This commit adds a "New Story" view which will be able to be used to
create a new story when it is completed. It allows all fields of a
story to be modified, unlike the existing "New Story" modal, and also
allow values for the stories fields to be specified in the query string
of the URL.

The New Story view also understands a `force_private` query string
parameter, which forces the creation of a private story by removing the
option from the user of unsetting the private flag.

The New Story view can be accessed at /#!/story/new, and allows various
query string parameters to be set to define the contents.

Change-Id: I8d4a2d8f195783daf6cfdbbd20998dad90c67c96