Remove bower and add yarn support
bower is deprecated and intended to be replaced. yarn is the recommended
replacement for both things installed via npm and things installed via
bower.

Move the bower resources into package.json and add the yarn.lock file.
Zuul has support for installing yarn if a yarn.lock file is found and
using it to install things.

Change-Id: I848f22dce6196b2df9de4b9f585e857e03cdb0d7