WIP Migrate to using webpack
This doesn't even come CLOSE to working yet. The biggest issue is
the LESS compilation step which does things with search paths in the
Gruntfile that can't realy be expressed in webpack or any of the less
plugins for webpack.

What we most likely want to do here, at least for now, is to replace the
Gruntfile with a very small Gulpfile that basically only knows how to
compile the LESS code. Then we can point the webpack config at the
compiled css output ... and we can put "gulp && webpack" in the npm
command entries.

It's hacky, but it's probably not TOO terrible to deal with for a
minute.

Then, people have expressed interest in shifting to SASS instead.
The webpack/bootstrap/saas support is a bit more robust and is what more
people seem to be doing, so as part of shifting from LESS to SASS we can
sort out making things be more webpack-native.

Change-Id: I6ec4cc04d7a4ba8e8c41169f5e236debc75de2be