Disable closing modals when backdrop is clicked
Closing modals when the backdrop is clicked can sometimes lead to
frustrating accidental data loss, for example when one is creating
a story and accidentally clicks outside the modal, losing everything
that has been typed.

This commit changes all the modals so that they can only be closed
using the cross or buttons in the modals. All modals are changed
to keep consistent behaviour across StoryBoard.

Change-Id: Id44ce43163c024eb5a6fe5744a3261848c0dcb74
Story: 2001553
Task: 6494