Make board and worklist icons unique
Currently boards and worklists share the task icon. This is
suboptimal, since the icon doesn't particularly suggest boards
or worklists at a glance.

This commit switches the board icon to one[0] which more resembles
a board, and the worklist item to a contained list to distinguish
from the tasks icon.

[0]: https://fontawesome.com/icons/columns?style=solid
[1]: https://fontawesome.com/icons/list-alt?style=regular

Change-Id: If89810896e25ce75fc7c36a22263ac1418ed8f69