Convert less to scss
This commit converts all our stylesheets from less to scss. Previously
Sass required ruby-compass, which wasn't easy to fit into our build
process. This is no longer the case so it makes sense to switch to
Sass/SCSS to enable us to streamline our build process and use some
of the more advanced features provided by Sass in the future.

Change-Id: I22fcbf4504d7756cd7ece41b0eca12985316db6b