Fix race when browsing stories by certain fields
When navigating to a view of search results, there is a race
condition where the API request for the browse happens twice,
once for the criteria which don't require API calls to create
useful Criteria for, and once for the full set.

This commit fixes this by waiting for the required extra API
calls to finish before updating the list of criteria.

Change-Id: I6af2f3a7185bd32a33afbb0fcafadd218e5d3567