Fix syntax highlighting
Syntax highlighting is currently broken because the npm package for
highlight.js doesn't contain a version of highlight.js built for use
in-browser. This commit fixes this by switching to using the latest
published version of the highlightjs package, which does contain a
version of highlight.js built for use in-browser.

Change-Id: Ie340d77e0423688d3b4b625a997c95e919fdd666