Make $scope.project into a promise in the detail view
Using a promise here allows us to use the id from $scope.project when
browsing for the stories in the project, rather than parsing it from
the $stateParams. This enables us to use non-numeric ids without having
to make a GET to find the numeric ID twice (once in each controller).

Change-Id: If80a3e423a700b1105a7e48d65eefb21ecef700f