Redesign the sidebar to neaten code and improve contrast
This commit completely rewrites the sidebar to use much cleaner
html, and simple to understand and maintain styling. The new
sidebar also has higher contrast for its contents and active
item. The scrolling behaviour still exists for some combinations
of permissions and screen size, but is made less painful since
only the overflowing part of the navbar now scrolls.

Change-Id: Ic9b71c9bfe03b8f324c8e36fa4dfd0266a2dcfcb