Don't display 404 errors when using the "Jump to..." box
Currently, typing an ID in the "Jump to..." box in the header allows the
user to jump directly to an object with a given ID. However, if a GET
for an object with that ID for the supported objects (currently Story,
Project, and Project Group) results in a 404, an error notification is
created, which gives a needless impression of something going wrong.

The lack of an object in the list of things with the entered ID is
sufficient indication that the object doesn't exist, without the extra
error message.

This commit adds a notification interceptor which stops 404 errors
raised between the "Jump to..." GET requests being sent and all of them
being resolved from being displayed to the user.

There is a small chance that this could cause useful 404 errors to be
missed, but the user would have to do something to cause it whilst the
other requests are still in-flight, which seems unlikely enough to be
acceptable to me.

Change-Id: I723d61479370e49b1e36ede0a118ad5d4464907f
Story: 2000466
Task: 2751