Make tags section more consistent
The tags section currently has a "button" which looks completely
different to all other buttons on the page, and also is the only
button which remains visible when logged out. The heading is also
slightly different to the other headings on the story detail view.

This commit fixes these inconsistencies to bring the tags section
in line with the rest of the story detail view.

Change-Id: Iceb368674849a08cc6e4456b8f7cfc283fc24a4f