Clear task title field on saving a new task
When adding many tasks consecutively, it is frustrating to have to clear
the task title field manually. It is unlikely that the user wants to
create many tasks with the same name.

This commit improves this by clearing the task title field when a new
task is saved, to remove the need to manually clear it.

Change-Id: I9ba08d9258de220a59a46c5b77163cba0b3d09c8