Remember scroll offset of lanes when refreshing board view
When the board object in the board detail view is reloaded (which is
quite a regular occurrence to help prevent stale data being used), the
scroll position of lanes is lost.

This is a minor inconvenience for small boards with only a few cards
and a short distance to scroll, but is frustrating for large boards
with many cards per lane, since it can be hard to find the old scroll
location. The issue is compounded by long load times of large boards
making it seem like the lane scrollbars reset with no apparent reason.

This commit solves this issue by adding support to remember the scroll
position of each lane when updating the board object, and re-scrolling
the newly rendered lanes to the correct position.

Change-Id: I75a6703b2f43bed8cdc4643f6a103f95e1fa61ae