Allow marking stories as security-related
This commit adds a checkbox when creating or editing stories which
allows them to be explicitly marked as security-related. It also rewords
the label for the privacy checkbox to avoid any confusion.

When creating a security-related story, privacy is enforced and the
checkbox disabled and replaced with a message explaining that. Public
security-related stories can be created by editing a story after the
fact to make it public, since there is no limitation on privacy when
editing a story. This is intended to make it difficult to accidentally
file a security story publically.

In order for the checkbox to actually be useful, this commit also adds
code which automates the addition of relevant security teams to the
story's permissions as tasks are added (or the security checkbox is
toggled).

Story: 2000568
Task: 29892
Task: 29893
Change-Id: I8a2c6e47f926aedcbe77878f2bc5991cd84f815d