Add a linter for our .scss files
This commit adds stylelint and some configuration for it. It also
updates the `lint` npm script to run stylelint on all our theme .scss
files.

There is also a large amount of churn in this commit in the .scss files
themselves, to bring them in line with the rules applied by stylelint.
The vast majority of this is standardising indentation to 2 spaces (we
previously used both 2 spaces and 4 spaces in different files).

Change-Id: I110a098f17c0fc2a514f0193426b96197c59282e