Fix post job by setting some variables
`node_version` is set to cause a version of node to be installed that
actually works to build StoryBoard on bionic.

The remaining issue with this job is that zuul currently uploads
everything in the dist/ directory to tarballs.openstack.org. The
output directory of our build script is also modified to not be "dist"
to stop this causing an untarballed webclient being uploaded to
tarballs.openstack.org. `create_tarball_directory` is set in .zuul.yaml
to reflect the new location.

Needed-By: I9f435c12df2b355ac9d735e9eda3f70cc396f1d9
Change-Id: Ia8bf1dc99ee18d31a6d269b243eade71fe519b7d