Fix post and release jobs by setting some variables
`node_version` is set to cause a version of node to be installed that
actually works to build StoryBoard on bionic.

The remaining issue with these jobs is that zuul currently uploads
everything in the dist/ directory to tarballs.openstack.org. The
output directory of our build script is also modified to not be "dist"
to stop this causing an untarballed webclient being uploaded to
tarballs.openstack.org. `create_tarball_directory` is set in .zuul.yaml
to reflect the new location.

Change-Id: Ia8bf1dc99ee18d31a6d269b243eade71fe519b7d