Load some config from /v1/systeminfo
This allows the API being used by a webclient to override things like
whether or not comments should be editable, based on the configuration
of the backend. This way the webclient can be more sensible about which
parts of the UI are usable, based on what the API claims to be able to
do rather than based on its own configuration file no matter what.

Old versions of the API don't return this information, so in that case
the webclient just uses whatever was in its config file.

Change-Id: I7bf6bfee6aba40e8ddb1e21bdee0f2737f8bd4d5