Converting a search query into a browse
At this point in my implementation, an email can be converted to an id
or the user may pass in their own id. Future commits will allow the
abilty for a user to have their query converted beyond emails.
The main change was having buildCriteriaMap return a promise
instead of a parameter object, and dealing with the functions
which expected that object.

Change-Id: Ia90582206cb6d1f648ef1e9468281473ef9bf88c