Enabling suggestions on all user fields
The big differences in this change consisted of creating seperate files
for each user type (subscriber, creator, assignee). This is to allow
different storyboard components to query on them as well instead of
being limited by the 'User' type that was previously used. The query is
still being done on the user type though it is split up within the
webclient. The new files are all the same besides minor details for now
though I chose this implementation incase the way we query changes in
the future. This commit also includes new icons for each user types.
It's worth noting that the browsing functionality usually makes the 
input into a tag, as a result of the routing in the board/worklists page
the browse -> tag functionality works only in the story component.

Story: 2005067
Task: 29637

Depends-On: https://review.openstack.org/#/c/644424/

Change-Id: I93e0a07dfb8e88c8484ebc87238fa5b5d2602f88
Signed-off-by: Malek Karray <malek.karray@windriver.com>