Allowing the user to choose what Columns are seen
This patch enables the user to choose what columns they can see
in the story menu. By checking or unchecking the item it will add
or remove the chosen column. The user also has the ability click on
a column to sort by that field. As of this patch, the frontend for
the 'Creator' column has been commented out because the backend
supports filtering by creator_id and not the creator's full name.
There has been talk of adding a full_name column to the story object
to support that sorting functionality.

Story: 2005413
Task: 30430

Change-Id: If36033b727786f3bb5bed1404f69530c33bcd7d8
Signed-off-by: Malek Karray <malek.karray@windriver.com>